refactor(routing):manage routin for more simple
