+++ frontend/src/app/precentation/pages/private/dashboard/dashboard.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { WrapperComponent } from '../../../../shared/container/wrapper.component';
+import { WrapperComponent } from '../../../shared/container/wrapper.component';
 
 @Component({
   selector: 'app-dashboard-presupuesto',
