+++ frontend/src/app/precentation/pages/private/layout.component.ts
@@ -1,10 +1,13 @@
 import { Component } from '@angular/core';
-import { NavComponent } from '../../../shared/nav/nav.component';
+import { NavComponent } from '../../shared/nav/nav.component';
 import { RouterOutlet } from '@angular/router';
+import { ModalComponent } from '../../shared/modal/modal.component';
+import { ToastComponent } from '../../../infraestructure/lib/toast/toast.component';
 
 @Component({
-  selector: 'app-unidad-layout',
+  selector: 'app-layout',
   template: `
+    <app-toast />
     <main class="grid grid-cols-[130px_1fr] h-screen">
       <app-nav class="h-full"></app-nav>
       <div class="bg-slate-200 h-full">
@@ -17,6 +20,6 @@ import { RouterOutlet } from '@angular/router';
       </div>
     </main>
   `,
-  imports: [NavComponent, RouterOutlet],
+  imports: [NavComponent, RouterOutlet, ToastComponent],
 })
-export class UnidadLayoutComponent {}
+export class LayoutComponent {}
