+++ frontend/src/app/precentation/pages/private/presupuestar/p_detail.component.ts
