--- frontend/src/app/precentation/pages/private/unidadEjecutora/dashborad/dashboard.component.ts
@@ -1,14 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  template: `<div class="">
-    <h1>Dashboad</h1>
-  </div>`,
-  selector: 'app-dashboard',
-  standalone: true,
-  imports: [],
-})
-export class DashboardComponent {
-  // This component can be used to display the dashboard for Unidad Ejecutora
-  // You can add any specific logic or data fetching here if needed
-}
