+++ frontend/src/app/precentation/pages/private/unidades/add.component.ts
@@ -1,8 +1,8 @@
 import { Component, inject } from '@angular/core';
-import { ContainerComponent } from '../../../../shared/container/container.component';
-import { CustomButtonComponent } from '../../../../shared/button/button.component';
-import { ModalComponent } from '../../../../shared/modal/modal.component';
-import { PanelService } from '../../../../../infraestructure/services/components/panel.service';
+import { ContainerComponent } from '../../../shared/container/container.component';
+import { CustomButtonComponent } from '../../../shared/button/button.component';
+import { ModalComponent } from '../../../shared/modal/modal.component';
+import { PanelService } from '../../../../infraestructure/services/components/panel.service';
 
 @Component({
   selector: 'app-add-unidades',
